+++ src/index.js
@@ -282,7 +282,7 @@ class Dayjs {
       weekdays, months, meridiem
     } = locale
     const getShort = (arr, index, full, length) => (
-      (arr && arr[index]) || full[index].substr(0, length)
+      (arr && (arr[index] || arr(this, str))) || full[index].substr(0, length)
     )
     const get$H = num => (
       Utils.s(this.$H % 12 || 12, num, '0')
