+++ test/locale/keys.test.js
@@ -35,7 +35,6 @@ Locale.forEach((locale) => {
     expect(weekdays).toEqual(expect.any(Array))
 
     if (weekdaysShort) expect(weekdaysShort).toEqual(expect.any(Array))
-    if (monthsShort) expect(monthsShort).toEqual(expect.any(Array))
     if (weekdaysMin) expect(weekdaysMin).toEqual(expect.any(Array))
     if (weekStart) expect(weekStart).toEqual(expect.any(Number))
 
@@ -44,6 +43,14 @@ Locale.forEach((locale) => {
       expect(months).toEqual(expect.any(Array))
     } else {
       expect(months(dayjs(), 'str')).toEqual(expect.any(String))
+    }
+    // monthsShort could be a function or array
+    if (monthsShort) {
+      if (Array.isArray(monthsShort)) {
+        expect(monthsShort).toEqual(expect.any(Array))
+      } else {
+        expect(monthsShort(dayjs(), 'str')).toEqual(expect.any(String))
+      }
     }
     // function pass date return string or number or null
     if (name !== 'en') { // en ordinal set in advancedFormat
